requirements.txt
# Web framework
Flask==3.1.2
flask-cors==5.0.1
flask-jwt-extended==4.7.1
flask-limiter==4.1.1
flask-talisman==1.1.0

# Database & migrations
SQLAlchemy==2.0.44
alembic==1.18.4

# Auth & security
bcrypt==4.3.0
python-dotenv==1.0.1

# Validation
marshmallow==3.26.2

# ML / audit engine
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
fairlearn==0.10.0
shap==0.46.0
joblib==1.5.2

# Testing
pytest==8.2.0
pytest-cov==5.0.0
